% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{add_hap_stats}
\alias{add_hap_stats}
\title{If the haplotyping gene has been specified, add haplotyping ratios to the genotype data frame, provided the ratios for that gene are suitable}
\usage{
add_hap_stats(genotype, hap_gene, haplo_details)
}
\arguments{
\item{genotype}{genotype data frame}

\item{hap_gene}{haplotyping gene for which ratios should be calculated}

\item{haplo_details}{Data structure created by create_haplo_details}
}
\value{
modified genotype with haplotyping ratios added
}
\description{
If the haplotyping gene has been specified, add haplotyping ratios to the genotype data frame, provided the ratios for that gene are suitable
}
