% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{calc_genotype}
\alias{calc_genotype}
\title{Build the genotype data required in the OGRDB genotype file}
\usage{
calc_genotype(segment, chain_type, s, ref_genes, inferred_seqs,
  genotype_db, hap_gene, haplo_details)
}
\arguments{
\item{segment}{one of V, D, J}

\item{chain_type}{one of H, L}

\item{s}{the input_sequences data frame}

\item{ref_genes}{named list of reference genes}

\item{inferred_seqs}{named list of novel gene sequences}

\item{genotype_db}{named list of gene sequences in the personalised genotype}

\item{hap_gene}{The haplotyping columns will be completed based on the usage of the two most frequent alleles of this gene. If NA, the column will be blank}

\item{haplo_details}{Data structure created by create_haplo_details}
}
\value{
A named list containing the following elements:
\tabular{ll}{
    input_sequences \tab     updated data frame, guaranteed to include mutation counts, which have been calculated if necessary \cr
    calculated_NC  \tab      a boolean, TRUE if mutation counts had to be calculated, FALSE otherwise \cr
    genotype \tab            data frame containing the information required for the OGRDB genotype file \cr
}
}
\description{
Build the genotype data required in the OGRDB genotype file
}
