% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{calc_haplo_details}
\alias{calc_haplo_details}
\title{Calculate the data used for haplotype analysis, showing allelic ratios calculated with various potential haplotyping genes}
\usage{
calc_haplo_details(segment, input_sequences)
}
\arguments{
\item{segment}{one of V, D, J}

\item{input_sequences}{the input_sequences data frame}
}
\value{
A named list containing the following elements:
\tabular{ll}{
    sa \tab           a variant of the input_sequences data frame, with fields for haplotyping analysis \cr
    a_genes \tab       list of potential haplotyping genes (if the segment under analysis is J, these are V genes, otherwise J genes) \cr
    a_props \tab       proportion of each allele of each potential haplotyping gene in the input sequences\cr
}
}
\description{
Calculate the data used for haplotype analysis, showing allelic ratios calculated with various potential haplotyping genes
}
