% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{gap_input_sequences}
\alias{gap_input_sequences}
\title{Check whether input sequences are gapped, and gap them if necessary}
\usage{
gap_input_sequences(input_sequences, inferred_seqs, ref_genes)
}
\arguments{
\item{input_sequences}{the input_sequences data frame}

\item{inferred_seqs}{named list of novel gene sequences}

\item{ref_genes}{named list of reference genes}
}
\value{
a revised input_sequences structure, guaranteed to contain gapped sequences in SEQUENCE_IMGT
}
\description{
Check whether input sequences are gapped, and gap them if necessary
}
