% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{generate_ogrdb_report}
\alias{generate_ogrdb_report}
\title{Generate OGRDB reports from specified files}
\usage{
generate_ogrdb_report(
  ref_filename,
  inferred_filename,
  species,
  filename,
  chain,
  hap_gene,
  segment,
  chain_type,
  plot_unmutated
)
}
\arguments{
\item{ref_filename}{Name of file containing IMGT-aligned reference genes in FASTA format}

\item{inferred_filename}{Name of file containing sequences of inferred novel alleles, or '-' if none}

\item{species}{Species name used in field 3 of the IMGT germline header with spaces omitted, if the reference file is from IMGT. Otherwise ''}

\item{filename}{Name of file containing annotated reads in AIRR, CHANGEO or IgDiscover format. The format is detected automatically}

\item{chain}{one of VH, VK, VL, D, JH, JK, JL}

\item{hap_gene}{The haplotyping columns will be completed based on the usage of the two most frequent alleles of this gene. If NA, the column will be blank}

\item{segment}{one of V, D, J}

\item{chain_type}{one of H, L}

\item{plot_unmutated}{Plot base composition using only unmutated sequences (V-chains only)}
}
\value{
Nothing
}
\description{
Generate OGRDB reports from specified files
}
