% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{genotype_statistics_cmd}
\alias{genotype_statistics_cmd}
\title{Collect parameters from the command line and use them to create a report and CSV file for OGRDB}
\usage{
genotype_statistics_cmd(test = F)
}
\arguments{
\item{test}{Run a simple preconfigured test (files must be in the current directory)}
}
\value{
Nothing
}
\description{
Collect parameters from the command line and use them to create a report and CSV file for OGRDB
}
