% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_barplot_grobs}
\alias{make_barplot_grobs}
\title{Create a barplot for each allele, showing number of reads distributed by mutation count}
\usage{
make_barplot_grobs(
  input_sequences,
  genotype_db,
  inferred_seqs,
  genotype,
  segment,
  calculated_NC
)
}
\arguments{
\item{input_sequences}{the input_sequences data frame}

\item{genotype_db}{named list of gene sequences in the personalised genotype}

\item{inferred_seqs}{named list of novel gene sequences}

\item{genotype}{data frame created by calc_genotype}

\item{segment}{one of V, D, J}

\item{calculated_NC}{a boolean, TRUE if mutation counts had to be calculated, FALSE otherwise}
}
\value{
list of grobs
}
\description{
Create a barplot for each allele, showing number of reads distributed by mutation count
}
\examples{
barplot_grobs = make_barplot_grobs(
                      example_rep$input_sequences,
                      example_rep$genotype_db,
                      example_rep$inferred_seqs,
                      example_rep$genotype,
                      'V',
                      example_rep$calculated_NC
               )

}
