% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{make_genotype_db}
\alias{make_genotype_db}
\title{Build a list of gene sequences that matches the personalised genotype}
\usage{
make_genotype_db(input_sequences, inferred_seqs, ref_genes)
}
\arguments{
\item{input_sequences}{the input_sequences data frame}

\item{inferred_seqs}{named list of novel gene sequences}

\item{ref_genes}{named list of reference genes}
}
\value{
named list of gene sequences
}
\description{
Build a list of gene sequences that matches the personalised genotype
}
