% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_haplo_grobs}
\alias{make_haplo_grobs}
\title{Create haplotyping plots}
\usage{
make_haplo_grobs(segment, haplo_details)
}
\arguments{
\item{segment}{one of V, D, J}

\item{haplo_details}{Data structure created by create_haplo_details}
}
\value{
named list containing the following elements:
\tabular{ll}{
    a_allele_plot \tab   plot showing allele usage for each potential haplotyping gene \cr
    haplo_grobs \tab     differential plot of allele usage for each usable haplotyping gene \cr
}
}
\description{
Create haplotyping plots
}
\examples{
haplo_grobs = make_haplo_grobs('V', example_rep$haplo_details)
}
