% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{read_inferred_sequences}
\alias{read_inferred_sequences}
\title{Read inferred sequences from a file. If ungapped, they will be gapped using the closest available reference gene}
\usage{
read_inferred_sequences(inferred_filename, segment, ref_genes)
}
\arguments{
\item{inferred_filename}{Name of file containing sequences of inferred novel alleles, or '-' if none}

\item{segment}{one of V, D, J}

\item{ref_genes}{Named list of reference sequences}
}
\value{
Named list of gapped inferred sequences
}
\description{
Read inferred sequences from a file. If ungapped, they will be gapped using the closest available reference gene
}
