% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{read_input_sequences}
\alias{read_input_sequences}
\title{Read annotated input sequences and convert format to CHANGEO-style, filling in as many gaps in what we get from the inference tool as we can}
\usage{
read_input_sequences(filename, segment, chain_type)
}
\arguments{
\item{filename}{Name of file containing annotated reads in AIRR, CHANGEO or IgDiscover format. The format is detected automatically}

\item{segment}{one of V, D, J}

\item{chain_type}{one of H, L}
}
\value{
Data Frame containing sequence annotations, with CHANGEO format headers.
}
\description{
Read annotated input sequences and convert format to CHANGEO-style, filling in as many gaps in what we get from the inference tool as we can
}
