% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{read_reference_genes}
\alias{read_reference_genes}
\title{Read the reference set from the specified file}
\usage{
read_reference_genes(ref_filename, species, chain, segment)
}
\arguments{
\item{ref_filename}{Name of file containing IMGT-aligned reference genes in FASTA format}

\item{species}{Species name used in field 3 of the IMGT germline header with spaces omitted, if the reference file is from IMGT. Otherwise ''}

\item{chain}{one of IGHV, IGKV, IGLV, IGHD, IGHJ, IGKJ, IGLJ, TRAV, TRAj, TRBV, TRBD, TRBJ, TRGV, TRGj, TRDV, TRDD, TRDJ}

\item{segment}{one of V, D, J}
}
\value{
Named list of gene sequences
}
\description{
Read the reference set from the specified file
}
