% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_statistics.R
\name{write_genotype_file}
\alias{write_genotype_file}
\title{Write the genotype file required by OGRDB}
\usage{
write_genotype_file(filename, segment, chain_type, genotype)
}
\arguments{
\item{filename}{name of file to create (csv)}

\item{segment}{one of V, D, J}

\item{chain_type}{one of H, L}

\item{genotype}{genotype data frame}
}
\value{
None
}
\description{
Write the genotype file required by OGRDB
}
\examples{
genotype_file = tempfile("ogrdb_genotype")
write_genotype_file(genotype_file, 'V', 'H', example_rep$genotype)
file.remove(genotype_file)
}
