% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_plot_file}
\alias{write_plot_file}
\title{Create the OGRDB style plot file}
\usage{
write_plot_file(
  filename,
  input_sequences,
  cdr3_dist_grobs,
  end_composition_grobs,
  whole_composition_grobs,
  triplet_composition_grobs,
  barplot_grobs,
  a_allele_plot,
  haplo_grobs,
  message
)
}
\arguments{
\item{filename}{name of file to create (pdf)}

\item{input_sequences}{the input_sequences data frame}

\item{cdr3_dist_grobs}{cdr3 length distribution grobs created by make_novel_base_grob}

\item{end_composition_grobs}{end composition grobs created by make_novel_base_grobs}

\item{whole_composition_grobs}{whole composition grobs created by make_novel_base_grobs}

\item{triplet_composition_grobs}{triplet composition grobs created by make_novel_base_grobs}

\item{barplot_grobs}{barplot grobs created by make_barplot_grons}

\item{a_allele_plot}{a_allele_plot grob created by make_haplo_grobs}

\item{haplo_grobs}{haplo_grobs created by make_haplo_grobs}

\item{message}{text message to display at end of report}
}
\value{
nothing
}
\description{
Create the OGRDB style plot file
}
