% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{genotype_statistics_test}
\alias{genotype_statistics_test}
\title{Generate OGRDB reports using test data}
\usage{
genotype_statistics_test(testdir, full = F)
}
\arguments{
\item{testdir}{Directory containing test data sets}

\item{full}{If false, run a single set. Otherwise run all sets we have.}
}
\value{
Nothing
}
\description{
Generate OGRDB reports using test data
}
